Move the code into a separate file
